@@ -35,9 +35,9 @@
     "pg": "^7.6.0",
     "prop-types": "^15.6.2",
     "query-string": "^6.2.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-chartjs-2": "^2.11.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^4.3.1",
     "semantic-ui-react": "^2.1.3",
     "styled-components": "^4.1.3"
