@@ -1,8 +1,9 @@
 import '@babel/polyfill';
-import ReactDOM from 'react-dom';
-// import App from './components/App';
+import { createRoot } from 'react-dom/client';
 import App from './components/App';
 
 require('./styles.css');
 
-ReactDOM.render(<App />, document.getElementById('app'));
+const container = document.getElementById('app');
+const root = createRoot(container);
+root.render(<App />);
